+++ src/app/student/components/student/student.component.ts
@@ -19,8 +19,17 @@ export class StudentComponent implements OnInit {
   }
 
   delete() {
-    this.studentService.delete(this.student.id).subscribe(e => {
-      this.isDelete = false;
-    }).unsubscribe();
+    this.studentService.getStudentsPracticeTickets(this.student.id).subscribe(tickets => {
+      if (tickets.length > 0) {
+        var confirm1 = confirm("У данного студента есть записи на практику, при его удалении заявки будут удалены, продолжить?");
+        if (!confirm1) {
+          return;
+        }
+      }
+      this.studentService.delete(this.student.id).subscribe(e => {
+        this.isDelete = true;
+      })
+    })
+
   }
 }
